move agend builder bindings, telemetry task runner class
