+++ x-pack/platform/plugins/shared/alerting_v2/server/index.ts
@@ -9,6 +9,7 @@ import { ContainerModule } from 'inversify';
 import type { PluginConfigDescriptor } from '@kbn/core/server';
 import type { PluginConfig } from './config';
 import { configSchema } from './config';
+import { bindAgentBuilder } from './setup/bind_agent_builder';
 import { bindContract } from './setup/bind_contract';
 import { bindOnSetup } from './setup/bind_on_setup';
 import { bindOnStart } from './setup/bind_on_start';
@@ -25,6 +26,7 @@ export const config: PluginConfigDescriptor<PluginConfig> = {
 
 export const module = new ContainerModule((options) => {
   bindOnSetup(options);
+  bindAgentBuilder(options);
   bindOnStart(options);
   bindContract(options);
   bindRoutes(options);
