+++ x-pack/platform/plugins/shared/alerting_v2/server/lib/usage/constants.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { IntervalSchedule } from '@kbn/task-manager-plugin/server';
+
+export const TELEMETRY_TASK_TYPE = 'alerting_v2_telemetry';
+export const TASK_ID = `AlertingV2-${TELEMETRY_TASK_TYPE}`;
+export const SCHEDULE: IntervalSchedule = { interval: '1d' };
