+++ x-pack/platform/plugins/shared/alerting_v2/server/lib/usage/task_definition.ts
@@ -5,27 +5,21 @@
  * 2.0.
  */
 
-import type { ElasticsearchClient, Logger } from '@kbn/core/server';
-import type { IntervalSchedule } from '@kbn/task-manager-plugin/server';
-import type { AlertingServerSetupDependencies } from '../../types';
-import { telemetryTaskRunner } from './task_runner';
+import type { AlertingTaskDefinition } from '../services/task_run_scope_service/create_task_runner';
+import { TELEMETRY_TASK_TYPE } from './constants';
+import { TelemetryTaskRunner } from './task_runner';
 import { stateSchemaByVersion } from './task_state';
 
-export const TELEMETRY_TASK_TYPE = 'alerting_v2_telemetry';
-export const TASK_ID = `AlertingV2-${TELEMETRY_TASK_TYPE}`;
-export const SCHEDULE: IntervalSchedule = { interval: '1d' };
-
-export function registerTelemetryTask(
-  logger: Logger,
-  taskManager: AlertingServerSetupDependencies['taskManager'],
-  getEsClient: () => ElasticsearchClient
-): void {
-  taskManager.registerTaskDefinitions({
-    [TELEMETRY_TASK_TYPE]: {
-      title: 'Alerting v2 snapshot telemetry fetch task',
-      timeout: '5m',
-      stateSchemaByVersion,
-      createTaskRunner: telemetryTaskRunner(logger, SCHEDULE, getEsClient),
-    },
-  });
-}
+/**
+ * Task definition for the snapshot telemetry fetch task.
+ * Bound to the TaskDefinition token and automatically registered with Task
+ * Manager on setup. Runs with the internal user (no fakeRequest required).
+ */
+export const TelemetryTaskDefinition: AlertingTaskDefinition<TelemetryTaskRunner> = {
+  taskType: TELEMETRY_TASK_TYPE,
+  title: 'Alerting v2 snapshot telemetry fetch task',
+  timeout: '5m',
+  stateSchemaByVersion,
+  taskRunnerClass: TelemetryTaskRunner,
+  requiresFakeRequest: false,
+};
